Refactored deleteImage and associated components

+++ app/actions/deleteImage.js
@@ -2,7 +2,7 @@
 
 import {createClient} from "@/utils/supabase/server";
 
-async function deleteImage(id, storageId) {
+export default async function deleteImage(id, storageId) {
 
 	const supabase = await createClient();
 
@@ -15,6 +15,9 @@ async function deleteImage(id, storageId) {
 	}
 
 	// delete image from storage
+	const storageSuccess = await deleteFromStorage(supabase, storageId);
+	if (!storageSuccess) return false;
+	/*
 	const {error: storageError} = await supabase
 		.storage
 		.from("images")
@@ -24,8 +27,12 @@ async function deleteImage(id, storageId) {
 		console.error("Storage Error: ", storageError);
 		return false;
 	}
+	*/
 
 	// delete image from database
+	const databaseSuccess = await deleteFromDatabase(supabase, id);
+	if (!databaseSuccess) return false;
+	/*
 	const {error: databaseError} = await supabase
 		.from("images")
 		.delete()
@@ -35,10 +42,41 @@ async function deleteImage(id, storageId) {
 		console.error("Database Error: ", databaseError);
 		return false;
 	}
+	*/
 
 	// successfully delete image
 	return true;
 
 }
 
-export default deleteImage;
+export async function deleteFromStorage(supabase, storageId) {
+
+	const {error} = await supabase
+		.storage
+		.from("images")
+		.remove([storageId]);
+	
+	if (error) {
+		console.error("Storage Error: ", error);
+		return false;
+	}
+
+	return true;
+
+}
+
+export async function deleteFromDatabase(supabase, id) {
+
+	const {error} = await supabase
+		.from("images")
+		.delete()
+		.eq("id", id);
+
+	if (error) {
+		console.error("Database Error: ", error);
+		return false;
+	}
+
+	return true;
+
+}
